requirements.txt
anthropic==0.43.1
speechrecognition==3.14.0
pyttsx3==2.98
python-dotenv==1.0.1
PyAudio==0.2.14